Initial work done to move the computation to the gpu using webgpu

+++ src/color.zig
@@ -3,18 +3,18 @@ const rtw = @import("rtweekend.zig");
 const std = rtw.std;
 const Interval = rtw.interval.Interval;
 
-pub const color: @Vector(3, f64) = undefined;
+pub const color: @Vector(3, f32) = undefined;
 
-pub fn linear_to_gamma(linear_component: f64) !f64 {
+pub fn linear_to_gamma(linear_component: f32) !f32 {
     if (linear_component > 0) return std.math.sqrt(linear_component) else return 0;
 }
 
-pub fn write_color(pixel_color: @Vector(3, f64)) !void {
+pub fn write_color(pixel_color: @Vector(3, f32)) !void {
     const stdout = std.io.getStdOut().writer();
 
-    var r: f64 = pixel_color[0];
-    var g: f64 = pixel_color[1];
-    var b: f64 = pixel_color[2];
+    var r: f32 = pixel_color[0];
+    var g: f32 = pixel_color[1];
+    var b: f32 = pixel_color[2];
 
     r = try linear_to_gamma(r);
     g = try linear_to_gamma(g);
